Remove singleton classes on CPK relations against Rails 4.1
* Extract CPK methods to its own module and extending instead of opening the singleton class.

* Fix `Marshal.dump`. Makes caching work again.
