+++ composite_primary_keys.gemspec
@@ -27,4 +27,5 @@ Gem::Specification.new do |s|
   # Dependencies
   s.required_ruby_version = '>= 1.9.3'
   s.add_dependency('activerecord', '=4.1.6')
+  s.add_development_dependency('rake', '~> 10.3.2')
 end
